karyon_p2p::bloom

Struct Bloom

pub struct Bloom {
    pub mandatory: u128,
    pub optional: u128,
}

Two 128-bit bloom filters (k=2 hashes) bundled into one wire type.

mandatory holds items the local node REQUIRES peers to also support.

optional holds items the local node would LIKE peers to support but doesn’t require.

Content-agnostic: items can be protocol ids, swarm keys, or any other identifier hashable as bytes. Discovery-layer hint, not

Fields
mandatory: u128optional: u128
Implementations

impl Bloom

pub fn empty() -> Self

Empty filter.

pub fn add_mandatory<I: AsRef<[u8]>>(&mut self, item: I)

Insert a mandatory item.

pub fn add_optional<I: AsRef<[u8]>>(&mut self, item: I)

Insert an optional item.

pub fn all(&self) -> u128

Union of mandatory + optional bits (every item the peer claims).

pub fn may_contain<I: AsRef<[u8]>>(&self, item: I) -> bool

True if this filter might contain item in either set.

pub fn covers_mandatory(&self, mine: &Self) -> bool

True if this peer’s full set of items covers every bit in mine.mandatory.

pub fn intersects_optional(&self, mine: &Self) -> bool

True if this peer’s full set of items shares at least one bit with mine.optional. Trivially false if mine.optional is empty.

pub fn is_empty(&self) -> bool

True if neither mandatory nor optional has any bit set.