Struct karyon_p2p::discovery::lookup::LookupService

pub struct LookupService {
    id: PeerID,
    table: Arc<Mutex<RoutingTable>>,
    listener: Arc<Listener>,
    connector: Arc<Connector>,
    outbound_slots: Arc<ConnectionSlots>,
    listen_endpoint: Option<RwLock<Endpoint>>,
    config: Arc<Config>,
    monitor: Arc<Monitor>,
}

Fields
id: PeerID

Peer’s ID

table: Arc<Mutex<RoutingTable>>

Routing Table

listener: Arc<Listener>

Listener

connector: Arc<Connector>

Connector

outbound_slots: Arc<ConnectionSlots>

Outbound slots.

listen_endpoint: Option<RwLock<Endpoint>>

Resolved listen endpoint

config: Arc<Config>

Holds the configuration for the P2P network.

monitor: Arc<Monitor>

Responsible for network and system monitoring.

Implementations

impl LookupService

pub fn new( key_pair: &KeyPair, id: &PeerID, table: Arc<Mutex<RoutingTable>>, config: Arc<Config>, monitor: Arc<Monitor>, ex: GlobalExecutor ) -> Self

Creates a new lookup service

pub async fn start(self: &Arc<Self>) -> Result<(), Error>

Start the lookup service.

pub async fn set_listen_endpoint(&self, resolved_endpoint: &Endpoint)

Set the resolved listen endpoint.

pub async fn shutdown(&self)

Shuts down the lookup service.

pub async fn start_lookup( &self, endpoint: &Endpoint, peer_id: Option<PeerID> ) -> Result<(), Error>

Starts iterative lookup and populate the routing table.

This method begins by generating a random peer ID and connecting to the provided endpoint. It then sends a FindPeer message containing the randomly generated peer ID. Upon receiving peers from the initial lookup, it starts connecting to these received peers and sends them a FindPeer message that contains our own peer ID.

async fn random_lookup( &self, endpoint: &Endpoint, peer_id: Option<PeerID>, random_peers: &mut Vec<PeerMsg> ) -> Result<(), Error>

Starts a random lookup

This will perfom lookup on a random generated PeerID

async fn self_lookup( &self, random_peers: &Vec<PeerMsg>, peer_buffer: &mut Vec<PeerMsg> )

Starts a self lookup

async fn connect( &self, endpoint: Endpoint, peer_id: Option<PeerID>, target_peer_id: &PeerID ) -> Result<Vec<PeerMsg>, Error>

Connects to the given endpoint and initiates a lookup process for the provided peer ID.

async fn handle_outbound( &self, io_codec: Codec, target_peer_id: &PeerID ) -> Result<Vec<PeerMsg>, Error>

Handles outbound connection

async fn start_listener(self: &Arc<Self>) -> Result<(), Error>

Start a listener.

async fn handle_inbound(self: &Arc<Self>, conn: Conn) -> Result<(), Error>

Handles inbound connection

async fn send_ping_msg(&self, io_codec: &Codec) -> Result<(), Error>

Sends a Ping msg and wait to receive the Pong message.

async fn send_pong_msg( &self, nonce: [u8; 32], io_codec: &Codec ) -> Result<(), Error>

Sends a Pong msg

async fn send_findpeer_msg( &self, io_codec: &Codec, peer_id: &PeerID ) -> Result<PeersMsg, Error>

Sends a FindPeer msg and wait to receivet the Peers msg.

async fn send_peers_msg( &self, peer_id: &PeerID, io_codec: &Codec ) -> Result<(), Error>

Sends a Peers msg.

async fn send_peer_msg( &self, io_codec: &Codec, endpoint: Endpoint ) -> Result<(), Error>

Sends a Peer msg.

async fn send_shutdown_msg(&self, io_codec: &Codec) -> Result<(), Error>

Sends a Shutdown msg.