Struct karyon_p2p::discovery::Discovery

pub struct Discovery {
    table: Arc<Mutex<RoutingTable>>,
    lookup_service: Arc<LookupService>,
    refresh_service: Arc<RefreshService>,
    connector: Arc<Connector>,
    listener: Arc<Listener>,
    conn_queue: Arc<ConnQueue>,
    pub(crate) inbound_slots: Arc<ConnectionSlots>,
    pub(crate) outbound_slots: Arc<ConnectionSlots>,
    task_group: TaskGroup<'static>,
    config: Arc<Config>,
}

Fields
table: Arc<Mutex<RoutingTable>>

Routing table

lookup_service: Arc<LookupService>

Lookup Service

refresh_service: Arc<RefreshService>

Refresh Service

connector: Arc<Connector>

Connector

listener: Arc<Listener>

Listener

conn_queue: Arc<ConnQueue>

Connection queue

inbound_slots: Arc<ConnectionSlots>

Inbound slots.

outbound_slots: Arc<ConnectionSlots>

Outbound slots.

task_group: TaskGroup<'static>

Managing spawned tasks.

config: Arc<Config>

Holds the configuration for the P2P network.

Implementations

impl Discovery

pub fn new( key_pair: &KeyPair, peer_id: &PeerID, conn_queue: Arc<ConnQueue>, config: Arc<Config>, monitor: Arc<Monitor>, ex: GlobalExecutor ) -> Arc<Discovery>

Creates a new Discovery

pub async fn start(self: &Arc<Self>) -> Result<(), Error>

Start the Discovery

pub async fn shutdown(&self)

Shuts down the discovery

async fn start_listener( self: &Arc<Self>, endpoint: &Endpoint ) -> Result<Endpoint, Error>

Start a listener and on success, return the resolved endpoint.

async fn connect_loop(self: Arc<Self>) -> Result<(), Error>

This method will attempt to connect to a peer in the routing table. If the routing table is empty, it will start the seeding process for finding new peers.

This will perform a backoff to prevent getting stuck in the loop if the seeding process couldn’t find any peers.

async fn connect(self: &Arc<Self>, endpoint: &Endpoint, pid: Option<PeerID>)

Connect to the given endpoint using the connector

async fn start_seeding(&self)

Starts seeding process.

This method randomly selects a peer from the routing table and attempts to connect to that peer for the initial lookup. If the routing table doesn’t have an available entry, it will connect to one of the provided bootstrap endpoints in the Config and initiate the lookup.

async fn random_entry(&self, entry_flag: u16) -> Option<Entry>

Returns a random entry from routing table.

async fn update_entry(&self, pid: &PeerID, entry_flag: u16)

Update the entry status