karyon_p2p

Struct Backend

pub struct Backend {
    config: Arc<Config>,
    key_pair: KeyPair,
    peer_id: PeerID,
    monitor: Arc<Monitor>,
    discovery: Arc<Discovery>,
    peer_pool: Arc<PeerPool>,
}

Backend serves as the central entry point for initiating and managing the P2P network.

Fields
config: Arc<Config>

The Configuration for the P2P network.

key_pair: KeyPair

Identity Key pair

peer_id: PeerID

Peer ID

monitor: Arc<Monitor>

Responsible for network and system monitoring.

discovery: Arc<Discovery>

Discovery instance.

peer_pool: Arc<PeerPool>

PeerPool instance.

Implementations

impl Backend

pub fn new(key_pair: &KeyPair, config: Config, ex: Executor) -> Arc<Backend>

Creates a new Backend.

pub async fn run(self: &Arc<Self>) -> Result<()>

Run the Backend, starting the PeerPool and Discovery instances.

pub async fn attach_protocol<P: Protocol>( &self, c: impl Fn(Arc<Peer>) -> Arc<dyn Protocol> + Send + Sync + 'static, ) -> Result<()>

Attach a custom protocol to the network

pub async fn peers(&self) -> usize

Returns the number of currently connected peers.

pub fn config(&self) -> Arc<Config>

Returns the Config.

pub fn peer_id(&self) -> &PeerID

Returns the PeerID.

pub fn key_pair(&self) -> &KeyPair

Returns the KeyPair.

pub async fn inbound_peers(&self) -> HashMap<PeerID, Endpoint>

Returns a map of inbound connected peers with their endpoints.

pub async fn outbound_peers(&self) -> HashMap<PeerID, Endpoint>

Returns a map of outbound connected peers with their endpoints.

pub fn monitor(&self) -> Arc<Monitor>

Returns the monitor to receive system events.

pub async fn shutdown(&self)

Shuts down the Backend.