Crate karyon_p2p

A lightweight, extensible, and customizable peer-to-peer (p2p) network stack.

Example

use std::sync::Arc;

use easy_parallel::Parallel;
use smol::{channel as smol_channel, future, Executor};

use karyon_core::crypto::{KeyPair, KeyPairType};
use karyon_p2p::{Backend, Config, PeerID};

let key_pair = KeyPair::generate(&KeyPairType::Ed25519);

// Create the configuration for the backend.
let mut config = Config::default();

// Create a new Executor
let ex = Arc::new(Executor::new());

// Create a new Backend
let backend = Backend::new(&key_pair, config, ex.clone());

let task = async {
    // Run the backend
    backend.run().await.unwrap();

    // ....

    // Shutdown the backend
    backend.shutdown().await;
};

future::block_on(ex.run(task));

Modules

backend 🔒

codec 🔒

config 🔒

connection 🔒

connector 🔒

discovery 🔒

error 🔒

listener 🔒

message 🔒

monitor

Responsible for network and system monitoring. Read More

peer 🔒

peer_pool 🔒

protocol

Defines the protocol trait. Read More

protocols 🔒

routing_table 🔒

slots 🔒

tls_config 🔒

version 🔒

Structs

Backend

Backend serves as the central entry point for initiating and managing the P2P network.

Config

the Configuration for the P2P network.

PeerID

Represents a unique identifier for a peer.

Version

Represents the network version and protocol version used in karyon p2p.

Enums

P2pError

Represents karyon’s p2p Error.

Type Aliases

ArcBackend

ArcPeer