Struct karyon_p2p::connector::Connector

pub struct Connector {
    key_pair: KeyPair,
    task_group: TaskGroup<'static>,
    connection_slots: Arc<ConnectionSlots>,
    max_retries: usize,
    enable_tls: bool,
    monitor: Arc<Monitor>,
}

Responsible for creating outbound connections with other peers.

Fields
key_pair: KeyPair

Identity Key pair

task_group: TaskGroup<'static>

Managing spawned tasks.

connection_slots: Arc<ConnectionSlots>

Manages available outbound slots.

max_retries: usize

The maximum number of retries allowed before successfully establishing a connection.

enable_tls: bool

Enables secure connection.

monitor: Arc<Monitor>

Responsible for network and system monitoring.

Implementations

impl Connector

pub fn new( key_pair: &KeyPair, max_retries: usize, connection_slots: Arc<ConnectionSlots>, enable_tls: bool, monitor: Arc<Monitor>, ex: GlobalExecutor ) -> Arc<Self>

Creates a new Connector

pub async fn shutdown(&self)

Shuts down the connector

pub async fn connect( &self, endpoint: &Endpoint, peer_id: &Option<PeerID> ) -> Result<Conn, Error>

Establish a connection to the specified endpoint. If the connection attempt fails, it performs a backoff and retries until the maximum allowed number of retries is exceeded. On a successful connection, it returns a Conn instance.

This method will block until it finds an available slot.

pub async fn connect_with_cback<Fut>( self: &Arc<Self>, endpoint: &Endpoint, peer_id: &Option<PeerID>, callback: impl FnOnce(Conn) -> Fut + Send + 'static ) -> Result<(), Error>where Fut: Future<Output = Result<(), Error>> + Send + 'static,

Establish a connection to the given endpoint. For each new connection, it invokes the provided callback, and pass the connection to the callback.

async fn dial( &self, endpoint: &Endpoint, peer_id: &Option<PeerID> ) -> Result<Conn, Error>