Module karyon_p2p::routing_table

Re-exports

pub use bucket::Bucket;

pub use bucket::BucketEntry;

pub use bucket::EntryStatusFlag;

pub use bucket::CONNECTED_ENTRY;

pub use bucket::DISCONNECTED_ENTRY;

pub use bucket::INCOMPATIBLE_ENTRY;

pub use bucket::PENDING_ENTRY;

pub use bucket::UNREACHABLE_ENTRY;

pub use bucket::UNSTABLE_ENTRY;

pub use entry::xor_distance;

pub use entry::Entry;

pub use entry::Key;

Modules

bucket 🔒

entry 🔒

Structs

RoutingTable

This is a modified version of the Kademlia Distributed Hash Table (DHT). https://en.wikipedia.org/wiki/Kademlia

Enums

AddEntryResult

Represents the possible result when adding a new entry.

Constants

DISTANCE_LIMIT 🔒

The distance limit for the closest buckets.

MAX_MATCHED_SUBNET_IN_BUCKET 🔒

The maximum number of matched subnets allowed within a single bucket.

MAX_MATCHED_SUBNET_IN_TABLE 🔒

The maximum number of matched subnets across the entire routing table.

TABLE_SIZE 🔒

The total number of buckets in the routing table.

Functions

subnet_match

Check if two addresses belong to the same subnet.