Module karyon_p2p::routing_table::bucket

Structs

Bucket

A Bucket represents a group of entries in the routing table.

BucketEntry

A BucketEntry represents a peer in the routing table.

Constants

ALL_ENTRY

BUCKET_SIZE

The number of entries that can be stored within a single bucket.

CONNECTED_ENTRY

The entry is connected.

DISCONNECTED_ENTRY

The entry is disconnected. This will increase the failure counter.

INCOMPATIBLE_ENTRY

The entry is incompatible. This entry will not contribute to an increase in failure attempts, instead, it will persist in the routing table for the lookup process and will only be removed in the presence of a new entry.

PENDING_ENTRY

The entry is ready to reconnect, meaning it has either been added and has no connection attempts, or it has been refreshed.

UNREACHABLE_ENTRY

The entry is unreachable. This will increase the failure counter.

UNSTABLE_ENTRY

The entry is unstable. This will increase the failure counter.

Type Aliases

EntryStatusFlag

BITFLAGS represent the status of an Entry within a bucket.